(Brian Jenkins/Richard Winslow): Added a full script path argument to flunk.sh and cleaned up some module.xml output details.
We changed the shell script so that you can pass either the -s/-t
combination to find your module script through the existing naming and
path conventions or you can specify a path to your script with the new
-S option.  We wanted to facilitate a different usage pattern in which
the script resides at the NetKernel modules directory level and uses a
different naming convention (e.g., urn.com.zynx.example.groovy).

We also updated the Module class so that its module.xml generation
features would avoid producing an empty (and invalid) mapper element
when there are no resource declarations that require one, and also to
produce an import from the execution space to the implementation space,
which should be part of the dependency injection pattern being
implemented here so that regular importers of the module can import only
the execution space.

We also had to add a JAR manifest artifact so that Java would
successfully find and run the Generate class' main method.