(Brian Jenkins/Richard Winslow): Added testlist generation option from command line and made modest changes to module.xml generated from test configuration.
You can now pass an addition, optional argument to the Generate.main
(command line) interface, following the script location, to specify
whether you want it to generate a module.xml file or a testlist.xml
file.  This argument takes the form of the word "module" or "testlist"
("module" being the default, if unspecified).

We also removed the unnecessary "version='2.0'" attribute from the
document element of the generated testlist.xml file (this appears to be
a copy/paste error) and changed the naming convention for test modules
to end with ":test" instead of ":tests" (this corresponds with our
conventions in the Galen project).

Finally, we fixed the sample minTime and maxTime assertions in
flunkTestDSL.groovy so that they are given integer arguments rather than
strings, as the testlist.xml generation was tripping up on the type
conversion, which didn't match the usage of these assertions in the unit
tests anyway.