(Alex Tatiyants/Richard Winslow): Introduced XML literal argument support, changed name of "pass_by" legal value, fixed a module-under-test name bug in TestList and fixed failures in GenerateTest and XUnitTest.
You can now pass an XML literal as an argument in a request declaration.
You specify the XML in the form of a closure that expects the "it"
parameter to be a MarkupBuilder and performs method calls on the builder
to produce the desired XML content.  The closure should be defined
outside the NetKernelBuilder.module (or .test) call, so that the builder
doesn't capture the closure and try to treat it like another layer of
language nesting.  See the new unit test for this feature for an
example.

Because the "method='as-string'" feature in NetKernel is a syntactic
sugar feature in NetKernel, while pass-by-value is a separate feature
that the as-string thing happens to incorporate, we changed the legal
value for the argument clause's "pass_by" attribute from "value" to
"argument-as-string" (here, the word "argument" is referring to argument
values from the endpoint grammar, since the as-string feature is
specifically designed to support the conversion of such input arguments
into string-type request arguments).

The test data file that the GenerateTest was using as part of its
expected output comparison had drifted apart from the input data
embedded in the test, so the test was failing.  There were also test
failures in XUnitTest because the buildModuleXML() function requires a
TestList to have module metadata.  We fixed these failures.

Finally, the TestList class was blindly assuming that the URI of the
module-under-test, which will point to a space with the suffix
convention ":impl" at the end, can just have the further suffix ":test"
added to it to form the name of the desired XUnit test module.  Instead,
the ":impl" should be removed from the end of the URI before turning it
into a test module URI.